# contracts/openapi/context-engine-v1.yaml
openapi: 3.1.0
info:
  title: Context Engine API
  version: 0.4.0
  description: Provider-neutral API draft for governed context spaces.
servers:
  - url: /v1
security:
  - bearerAuth: []
paths:
  /spaces:
    get:
      operationId: listContextSpaces
      responses:
        "200":
          description: Visible context spaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContextSpace"
    post:
      operationId: createContextSpace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContextSpace"
      responses:
        "201":
          description: Context space created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContextSpace"
        "400":
          $ref: "#/components/responses/Error"
  /spaces/{spaceId}:
    parameters:
      - $ref: "#/components/parameters/SpaceId"
    get:
      operationId: getContextSpace
      responses:
        "200":
          description: Visible context space
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContextSpace"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateContextSpace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContextSpace"
      responses:
        "200":
          description: Updated context space
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContextSpace"
    delete:
      operationId: deleteContextSpace
      responses:
        "202":
          $ref: "#/components/responses/JobAccepted"
  /spaces/{spaceId}/sources:
    parameters:
      - $ref: "#/components/parameters/SpaceId"
    get:
      operationId: listSources
      responses:
        "200":
          description: Registered sources in a visible space
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Source"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    post:
      operationId: registerSource
      description: Requires source.manage or space.manage on the space.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterSource"
      responses:
        "201":
          description: Source registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /sources/{sourceId}:
    parameters:
      - $ref: "#/components/parameters/SourceId"
    get:
      operationId: getSource
      responses:
        "200":
          description: Visible source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateSource
      description: Rename, pause or resume, or remap a source; requires source.manage or space.manage.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSource"
      responses:
        "200":
          description: Updated source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /sources/{sourceId}/checkpoints:
    parameters:
      - $ref: "#/components/parameters/SourceId"
    get:
      operationId: getSourceCheckpoint
      responses:
        "200":
          description: Current connector cursor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkpoint"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    put:
      operationId: putSourceCheckpoint
      description: Store the connector cursor; requires ingest.write on the source.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [cursor]
              properties:
                cursor:
                  type: string
                  minLength: 1
                  maxLength: 4000
      responses:
        "200":
          description: Stored connector cursor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkpoint"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /sources/{sourceId}/uploads:
    post:
      operationId: stageUpload
      description: >-
        Stage content bytes for a later ingestion event. The response uploadId becomes the
        event's contentRef and its contentHash and contentType must match. Requires
        ingest.write on the source.
      parameters:
        - $ref: "#/components/parameters/SourceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Content staged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Upload"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "413":
          $ref: "#/components/responses/Error"
        "415":
          $ref: "#/components/responses/Error"
  /sources/{sourceId}/records/{recordId}:
    get:
      operationId: getRecordStatus
      description: Record lifecycle without content, audiences, or backend details.
      parameters:
        - $ref: "#/components/parameters/SourceId"
        - name: recordId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 500
      responses:
        "200":
          description: Current record state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordStatus"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /sources/{sourceId}/jobs:
    get:
      operationId: listSourceJobs
      description: Deliveries for a source, for dead-letter inspection; requires delivery or management rights.
      parameters:
        - $ref: "#/components/parameters/SourceId"
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum: [queued, running, succeeded, failed]
      responses:
        "200":
          description: Jobs newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /ingestions:
    post:
      operationId: acceptIngestion
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../schemas/ingestion-event.schema.json"
      responses:
        "202":
          $ref: "#/components/responses/JobAccepted"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /jobs/{jobId}:
    get:
      operationId: getJob
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Job state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          $ref: "#/components/responses/Error"
...
  /auth/me:
    get:
      operationId: getCurrentPrincipal
      description: The calling principal as the engine resolved it from the credential.
      responses:
        "200":
          description: Current principal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Principal"
        "401":
          $ref: "#/components/responses/Error"
  /auth/permissions:
    get:
      operationId: listEffectivePermissions
      parameters:
        - name: resourceId
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Identifier"
      responses:
        "200":
          description: Effective engine actions of the caller on a visible resource
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [resourceId, actions]
                properties:
                  resourceId:
                    $ref: "#/components/schemas/Identifier"
                  actions:
                    type: array
                    uniqueItems: true
                    items:
                      $ref: "#/components/schemas/Action"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /resources/{resourceId}/grants:
    get:
      operationId: listGrants
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Grants on the resource; requires access.manage
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grant"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /resources/{resourceId}/grants/{grantId}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
      - name: grantId
        in: path
        required: true
        schema:
          type: string
          pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,199}$"
    put:
      operationId: putGrant
      description: >-
        Create or replace a grant. The acting principal comes from the credential; the body
        names only the grant target and actions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutGrant"
      responses:
        "200":
          description: Grant stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grant"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteGrant
      responses:
        "204":
          description: Grant removed
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SpaceId:
      name: spaceId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
    SourceId:
      name: sourceId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 200
  responses:
    Error:
      description: Stable engine error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    JobAccepted:
      description: Asynchronous work accepted
      headers:
        Location:
          schema:
            type: string
          description: Job status URL
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JobAccepted"
...
  schemas:
...
    ContextSpace:
      type: object
      additionalProperties: false
      required: [id, name, state, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/Identifier"
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 1000
        state:
          type: string
          enum: [provisioning, ready, failed, deleting]
        createdAt:
          type: string
          format: date-time
    CreateContextSpace:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 1000
    Source:
      type: object
      additionalProperties: false
      required: [id, spaceId, name, type, state, versionOrdering, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/Identifier"
        spaceId:
          $ref: "#/components/schemas/Identifier"
        name:
          type: string
          minLength: 1
          maxLength: 120
        type:
          type: string
          minLength: 1
          maxLength: 60
        state:
          type: string
          enum: [ready, paused, failed]
        versionOrdering:
          $ref: "#/components/schemas/VersionOrdering"
        createdAt:
          type: string
          format: date-time
...
    RegisterSource:
      type: object
      additionalProperties: false
      required: [name, type, audienceMapping]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        type:
          type: string
          minLength: 1
          maxLength: 60
        audienceMapping:
          $ref: "#/components/schemas/AudienceMapping"
        versionOrdering:
          $ref: "#/components/schemas/VersionOrdering"
    UpdateSource:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        state:
          type: string
          enum: [ready, paused]
        audienceMapping:
          $ref: "#/components/schemas/AudienceMapping"
    Checkpoint:
      type: object
      additionalProperties: false
      required: [sourceId, cursor, updatedAt]
      properties:
        sourceId:
          $ref: "#/components/schemas/Identifier"
        cursor:
          type: string
        updatedAt:
          type: string
          format: date-time
    Upload:
      type: object
      additionalProperties: false
      required: [uploadId, sourceId, contentType, sizeBytes, contentHash, expiresAt]
      properties:
        uploadId:
          $ref: "#/components/schemas/Identifier"
        sourceId:
          $ref: "#/components/schemas/Identifier"
        contentType:
          type: string
        sizeBytes:
          type: integer
          minimum: 1
        contentHash:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        expiresAt:
          type: string
          format: date-time
    RecordStatus:
      type: object
      additionalProperties: false
      required: [spaceId, sourceId, recordId, state, currentVersion, sourceAclVersion, updatedAt]
      properties:
        spaceId:
          $ref: "#/components/schemas/Identifier"
        sourceId:
          $ref: "#/components/schemas/Identifier"
        recordId:
          type: string
        state:
          type: string
          enum: [active, deleted, quarantined]
        currentVersion:
          type: string
        sourceAclVersion:
          type: string
        contentHash:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        quarantineReason:
          type: string
        updatedAt:
          type: string
          format: date-time
...